+++ src/games/chlcc/albummenu.cpp
@@ -31,7 +31,7 @@ void AlbumMenu::CgOnClick(Widgets::Button* target) {
   SaveSystem::GetEVStatus(target->Id, &total, &viewed);
 
   ShowCgViewer = true;
-  CgViewerWidget->LoadCgSprites(target->Id, "bg",
+  CgViewerWidget->LoadCgSprites((size_t)target->Id, "bg",
                                 Profile::SaveSystem::AlbumData[target->Id]);
 }
