+++ src/ui/widgets/label.h
@@ -50,7 +50,7 @@ class Label : public Widget {
     Bounds = {};
   }
 
-  int GetTextLength() { return Text.size(); }
+  size_t GetTextLength() { return Text.size(); }
   float GetFontSize() { return FontSize; }
 
   float OutlineAlpha = 1.0f;
